Moved the files inside the folder to see if the documentation is generated on the GitHub docs.
