+++ tests/test_data_loading.py
@@ -1,4 +1,4 @@
-from data_loading import load_data_from_h5, load_reference_molecule
+from compare_geoms.data_loading import load_data_from_h5, load_reference_molecule
 
 sample_h5_file = 'output_9953.h5'
 sample_ref_molecule = '264_noise00.xyz'
